+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 ==========
+## Version 6.0.16
+_2024-04-18_
+
+- `viewMeta` Exposed for bridging with cross-platform agents
+
 ## Version 6.0.15
 _2024-02-28_